+++ gr-pdu/lib/pdu_split_impl.cc
@@ -28,11 +28,10 @@ pdu_split_impl::pdu_split_impl(const bool pass_empty_data)
     : gr::block("pdu_split", io_signature::make(0, 0, 0), io_signature::make(0, 0, 0)),
       d_pass_empty_data(pass_empty_data)
 {
-    message_port_register_in(PMTCONSTSTR__pdus());
-    set_msg_handler(PMTCONSTSTR__pdus(),
-                    [this](pmt::pmt_t msg) { this->handle_pdu(msg); });
-    message_port_register_out(PMTCONSTSTR__dict());
-    message_port_register_out(PMTCONSTSTR__data());
+    message_port_register_in(ports::pdus());
+    set_msg_handler(ports::pdus(), [this](pmt::pmt_t msg) { this->handle_pdu(msg); });
+    message_port_register_out(ports::dict());
+    message_port_register_out(ports::data());
 }
 
 pdu_split_impl::~pdu_split_impl() {}
@@ -49,11 +48,11 @@ void pdu_split_impl::handle_pdu(pmt::pmt_t pdu)
     pmt::pmt_t data = pmt::cdr(pdu);
 
     if ((!pmt::equal(meta, pmt::PMT_NIL)) | d_pass_empty_data) {
-        message_port_pub(PMTCONSTSTR__dict(), meta);
+        message_port_pub(ports::dict(), meta);
     }
 
     if (pmt::length(data) | d_pass_empty_data) {
-        message_port_pub(PMTCONSTSTR__data(), data);
+        message_port_pub(ports::data(), data);
     }
 }