+++ gr-pdu/lib/random_pdu_impl.cc
@@ -36,7 +36,7 @@ random_pdu_impl::random_pdu_impl(int min_items,
       d_mask(byte_mask),
       d_length_modulo(length_modulo)
 {
-    message_port_register_out(PMTCONSTSTR__pdus());
+    message_port_register_out(ports::pdus());
     message_port_register_in(pmt::mp("generate"));
     set_msg_handler(pmt::mp("generate"),
                     [this](pmt::pmt_t msg) { this->generate_pdu(msg); });
@@ -67,7 +67,7 @@ void random_pdu_impl::output_random()
     pmt::pmt_t vecpmt(pmt::make_blob(&vec[0], len));
     pmt::pmt_t pdu(pmt::cons(pmt::PMT_NIL, vecpmt));
 
-    message_port_pub(PMTCONSTSTR__pdus(), pdu);
+    message_port_pub(ports::pdus(), pdu);
 }
 
 } /* namespace pdu */