+++ gr-pdu/lib/tagged_stream_to_pdu_impl.h
@@ -18,13 +18,14 @@ namespace pdu {
 
 class PDU_API tagged_stream_to_pdu_impl : public tagged_stream_to_pdu
 {
-    const pdu::vector_type d_type;
+    const pdu::types::vector_type d_type;
     pmt::pmt_t d_pdu_meta;
     pmt::pmt_t d_pdu_vector;
     std::vector<tag_t> d_tags;
 
 public:
-    tagged_stream_to_pdu_impl(pdu::vector_type type, const std::string& lengthtagname);
+    tagged_stream_to_pdu_impl(pdu::types::vector_type type,
+                              const std::string& lengthtagname);
 
     int work(int noutput_items,
              gr_vector_int& ninput_items,