+++ gr-pdu/python/pdu/bindings/pdu_to_tagged_stream_python.cc
@@ -14,7 +14,7 @@
 /* BINDTOOL_GEN_AUTOMATIC(0)                                                       */
 /* BINDTOOL_USE_PYGCCXML(0)                                                        */
 /* BINDTOOL_HEADER_FILE(pdu_to_tagged_stream.h)                                        */
-/* BINDTOOL_HEADER_FILE_HASH(dd83e789baa1353ca99e1cf560e83b01)                     */
+/* BINDTOOL_HEADER_FILE_HASH(3a93b879f5a4b6786b1ec4598e9a8786)                     */
 /***********************************************************************************/
 
 #include <pybind11/complex.h>